backend-flask/requirements.txt
flask
flask-cors

aws-xray-sdk

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests

watchtower


blinker
rollbar

cognitojwt

psycopg[pool]
psycopg[binary]

